changed some things to size_t to reduce warnings on MSVC x64
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5492 d7cf8633-e32d-0410-b094-e92efae38249

+++ zone.c
@@ -26,7 +26,7 @@ cvar_t developer_memorydebug = {0, "developer_memorydebug", "0"};
 
 mempool_t *poolchain = NULL;
 
-void *_Mem_Alloc(mempool_t *pool, int size, const char *filename, int fileline)
+void *_Mem_Alloc(mempool_t *pool, size_t size, const char *filename, int fileline)
 {
 #if MEMCLUMPING
 	int i, j, k, needed, endbit, largest;
